refactor: ajustando atividades 2,3 e 4
