+++ E02/src/bdn.c
