+++ E02/src/bdu.c
