+++ E02/src/cdb.c
