+++ E02/src/tecnicas.h
