+++ E04/src/automacao.sh
